cours_service/Cours/requirements.txt
djangorestframework
PyPDF2